refactor: remove redundant evolve code
