+++ pycascades/__init__.py
@@ -8,12 +8,7 @@ __credits__ = 'Potsdam Institute for Climate Impact Research'
 from . import core, gen, utils, earth_system, amazon
 
 from pycascades.core.tipping_element import cusp, hopf, realistic_cusp
-from pycascades.core.tipping_element_economic import economic_logistic
 from pycascades.core.tipping_network import tipping_network
-from pycascades.core.tipping_network_economic import tipping_network_economic
 from pycascades.core.coupling import linear_coupling, cusp_to_hopf, hopf_x_to_cusp, hopf_x_to_hopf, hopf_y_to_cusp, hopf_y_to_hopf
-from pycascades.core.coupling_economic import linear_coupling as linear_coupling_economic
 from pycascades.core.evolve import evolve
-from pycascades.core.evolve_sde import evolve as evolve_sde
-from pycascades.core.evolve_economic import evolve as evolve_economic
 from pycascades.utils import plotter
